WIP on splitting search to a web worker
